+++ pycontrails/models/cocip/contrail_properties.py
@@ -1214,7 +1214,7 @@ def plume_temporal_evolution(
     """
     # Convert dt to seconds value and use dtype of other variables
     dtype = np.result_type(width_t1, depth_t1, sigma_yz_t1, dsn_dz_t1, diffuse_h_t1, diffuse_v_t1)
-    dt_s = units.dt_to_float_seconds(dt, dtype)
+    dt_s = units.dt_to_seconds(dt, dtype)
 
     sigma_yy = 0.125 * width_t1**2
     sigma_zz = 0.125 * depth_t1**2
@@ -1404,7 +1404,7 @@ def new_ice_particle_number(
     """
     # Convert dt to seconds value and use dtype of other variables
     dtype = np.result_type(n_ice_per_m_t1, dn_dt_agg, dn_dt_turb, seg_ratio)
-    dt_s = units.dt_to_float_seconds(dt, dtype)
+    dt_s = units.dt_to_seconds(dt, dtype)
 
     n_ice_per_m_t1 = np.maximum(n_ice_per_m_t1, 0.0)
 
@@ -1525,5 +1525,5 @@ def mean_energy_flux_per_m(
     --------
     :func:`mean_radiative_flux_per_m`
     """
-    dt_s = units.dt_to_float_seconds(dt, rad_flux_per_m.dtype)
+    dt_s = units.dt_to_seconds(dt, rad_flux_per_m.dtype)
     return rad_flux_per_m * dt_s